requirements.txt
fastapi==0.135.3
pydantic[email]
uvicorn==0.44.0
supabase==2.28.3
structlog==25.5.0
requests==2.32.4
python-dotenv
tzdata
apscheduler==3.11.0
pytz
httpx
pytest
arq
redis